requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
psycopg2-binary
google-auth
google-auth-oauthlib
google-api-python-client
gspread
argon2-cffi
aiofiles
itsdangerous
python-multipart
Pillow
fpdf
segno
cloudinary
